pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "multiplex_pipeline"
dynamic = ["version"]
description = "A pipeline for analyzing multiplexed Cell Dive data"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Kasia Kedziora", email = "kasia.kedziora@pitt.edu" }
]
dependencies = [
    "anndata>=0.11.4",
    "dask[array]>=2024.4",
    "globus-sdk>= 3.59.0",
    "imagecodecs>=2024.9.22",
    "loguru>=0.7.3",
    "numpy>=2.0",
    "opencv-python>=4.11",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "shapely>=2.1.1",
    "spatialdata>=0.4",
    "tifffile>=2025.1.10,<2025.5.21",
    "zarr>=2.18.7",
    "fsspec>=2024.6.1",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

[project.optional-dependencies]
segmentation-gpu = [
    "cellpose>=4.0",
    "imagecodecs>=2024.6.1",
    "instanseg-torch>=0.0.9",
    "torch",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]


[tool.setuptools_scm]
write_to = "src/multiplex_pipeline/_version.py"


[tool.black]
line-length = 79
target-version = ['py38', 'py39', 'py310']

[tool.ruff]
line-length = 79
lint.select = [
    "E", "F", "W", #flake8
    "UP", # pyupgrade
    "I", # isort
    "BLE", # flake8-blind-exception
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
lint.ignore = [
    "E501", # line too long. let black handle this
    "UP006", "UP007", # type annotation. As using magicgui require runtime type annotation then we disable this.
    "SIM117", # flake8-simplify - some of merged with statements are not looking great with black, reanble after drop python 3.9
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*vendored*",
    "*_vendor*",
    "*.ipynb",
]
